Move get_pending_transfers_tree from test.utils to utils
We use it in a tool in the main repo now too.
